% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{asymmetry}
\alias{asymmetry}
\title{Asymmetry scores}
\usage{
asymmetry(sm, mfconfig)
}
\arguments{
\item{sm}{a face matrix}

\item{mfconfig}{face config info including left right pairs}
}
\value{
scores for each marker
}
\description{
Measures of asymmetry in object containing some matched pairs
}
\details{
Face is reflected and the ordinary Procrustes applied to match up
with distances calculated between markers
}
