% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{averecface}
\alias{averecface}
\title{Average faces and rotate so head is upright}
\usage{
averecface(shapem, mconfig)
}
\arguments{
\item{shapem}{An array of facial shapes}

\item{mconfig}{configuration info}
}
\value{
A facial shape matrix rotated to have minimum variance in frontal and lateral directions
}
\description{
Average faces and rotate so head is upright
}

