% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{cleansm}
\alias{cleansm}
\title{Process a facial motion to fill-in missing values and report diagnostics}
\usage{
cleansm(sm, MISSMAX = 50)
}
\arguments{
\item{sm}{a facial motion}

\item{MISSMAX}{maximum allowable number of missing cases in a marker (default=50)}
}
\value{
a list containing a facial motion with missing values filled in, the number of
missing cases in each marker, the max jump in a marker between frames, which cases
are all missing and a matrix of where the missing values occur
}
\description{
Process a facial motion to fill-in missing values and report diagnostics
}
