% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{constructfaceseq}
\alias{constructfaceseq}
\title{Array of faces constructed from geodesic and profile information}
\usage{
constructfaceseq(initshap, farshap, prof)
}
\arguments{
\item{initshap}{shape matrix with starting posture}

\item{farshap}{shape matrix with extreme posture}

\item{profile}{vector of value in [0,1] for convex combination of postures}
}
\value{
array where each frame is a shape matrix. Number of frames is the length of the profile vector.
}
\description{
Array of faces constructed from geodesic and profile information
}
