% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{constructscorefaceseq}
\alias{constructscorefaceseq}
\title{Array of faces constructed from geodesic and score informationTitle}
\usage{
constructscorefaceseq(initshap, farshap, score)
}
\arguments{
\item{initshap}{initial shape}

\item{farshap}{extremal shape}

\item{score}{score profile}
}
\value{
facial motion sequence with same number frames as the score
}
\description{
Array of faces constructed from geodesic and score informationTitle
}
