% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{extmfconfig}
\alias{extmfconfig}
\title{Extract information about the motion capture file}
\usage{
extmfconfig(fname, nskip = 5)
}
\arguments{
\item{fname}{A name of a typical motion capture file}

\item{nskip}{Number of lines at top containing the info}
}
\value{
list containing number of frames, number of markers, names of the
  markers, the markers which lie on the midline of the face and the
  coordinate directions as (lateral, vertical, frontal)
}
\description{
Extract information about the motion capture file
}
