% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{faceframe}
\alias{faceframe}
\title{Construct a single frame suitable for movie construction}
\usage{
faceframe(fc1, mfconfig, fc2 = NULL, rang = NULL, box = FALSE,
  header = FALSE)
}
\arguments{
\item{fc1}{A face}

\item{mfconfig}{face configuration}

\item{fc2}{Another face (optional)}

\item{box}{use boxes (default is FALSE)}

\item{header}{optional title for plot}
}
\value{
nothing - a plot is created
}
\description{
Construct a single frame suitable for movie construction
}

