% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{faceGPA}
\alias{faceGPA}
\title{Generalized Procrustes Analysis}
\usage{
faceGPA(A, itmax = 5, pcaoutput = FALSE)
}
\arguments{
\item{A}{a 3D array arranged as markers x dimensions x faces}

\item{itmax}{maximum number of iterations (5 by default)}
}
\value{
a list containing the rotated array of faces, the mean face and the centroid sizes of the faces
}
\description{
Generalized Procrustes Analysis
}
