% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{faceimpute}
\alias{faceimpute}
\title{Impute missing values in a face file}
\usage{
faceimpute(mface, tface, verbose = FALSE)
}
\arguments{
\item{mface}{a face file with missing values}

\item{tface}{a face file for training}

\item{verbose}{logical value indicating whether to print out missing value info}
}
\value{
a face file with missing values imputed
}
\description{
Impute missing values in a face file
}
