% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{facemovie}
\alias{facemovie}
\title{Make a facial motion movie}
\usage{
facemovie(shapem1, mfconfig, shapem2 = NULL, everynth = 6, adjust = TRUE,
  moviename = "movie.mov", header = NULL)
}
\arguments{
\item{shapem1}{A shape motion}

\item{mfconfig}{motion config}

\item{shapem2}{another shape motion (optional)}

\item{everynth}{sample every nth frame (default is 6)}

\item{adjust}{adjust second motion to start from same place as first motion}

\item{moviename}{name of the movie (default is "movie.mov")}

\item{header}{title displayed on movie}
}
\value{
nothing - find movie file in "animations" directory
}
\description{
Make a facial motion movie
}

