% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{facenorm}
\alias{facenorm}
\title{Compute size of face}
\usage{
facenorm(f)
}
\arguments{
\item{f}{the face}
}
\value{
the size of the face
}
\description{
Average distance of markers from the origin
}
\details{
Assumes the face has been centered
}
