% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{faceOPA}
\alias{faceOPA}
\title{Ordinary Procrustes Analysis}
\usage{
faceOPA(A, B)
}
\arguments{
\item{A}{a face matrix}

\item{B}{another face matrix with the same dimension}
}
\value{
a list with the following components:
\item{Ahat}{centered A}
\item{Bhat}{B centered and rotated onto A}
\item{OSS}{sum of squares for the difference}
\item{rmsd}{RMS difference = sqrt(OSS/(no. of markers))}
}
\description{
Fit one face onto another
}
\details{
Face A is centered, Face B is centered then rotated to obtain the best Procrustes fit
}
