% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{first_PCA_traj}
\alias{first_PCA_traj}
\title{extract the first PCA, select start and max, return initial and max frames}
\usage{
first_PCA_traj(sm)
}
\arguments{
\item{sm}{The motion data array}
}
\value{
list containing the first PC score trajectory, the frame number where the maximum is obtained,
the initial frame, the the frame at the max, the geodesic initial frame and the geodesic max frame
}
\description{
extract the first PCA, select start and max, return initial and max frames
}
