% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{misfixlin}
\alias{misfixlin}
\title{Fill in missing values using linear interpolation}
\usage{
misfixlin(x)
}
\arguments{
\item{x}{1D motion trace}
}
\value{
1D motion trace
}
\description{
Fill in missing values using linear interpolation
}

