% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{multifaceplot}
\alias{multifaceplot}
\title{Plot an array of shapes with a different color for each marker}
\usage{
multifaceplot(fm, mfconfig, view = c("front", "side", "top"), main = NULL,
  subset = NULL)
}
\arguments{
\item{fm}{array markers by directions by cases}

\item{mfconfig}{configuration information}

\item{view}{front, side or top}

\item{main}{optional title}
}
\value{
nothing
}
\description{
Plot an array of shapes with a different color for each marker
}
