% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{parproc}
\alias{parproc}
\title{Partial Procrustes rotation}
\usage{
parproc(A, B)
}
\arguments{
\item{A}{a face matrix}

\item{B}{a face matrix}
}
\value{
the rotated B
}
\description{
rotates B onto A
}
\details{
Assumes that A and B have both been centered (and preferably scaled)
}
