% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{plotface}
\alias{plotface}
\title{Plot a face or a pair of faces}
\usage{
plotface(fm, mfconfig, fm2 = NULL, view = c("front", "side", "top"),
  num = TRUE, main = "", markerlabs = NULL, dirm = c("arrows", "points",
  "segments"), subset = 1:nrow(fm))
}
\arguments{
\item{fm}{A face matrix}

\item{mfconfig}{motion capture config file}

\item{fm2}{Another face matrix (optional)}

\item{view}{direction of the 2D view}

\item{num}{Marker numbers, rather than points, to be displayed}

\item{main}{option title}

\item{markerlabs}{optional marker labels}

\item{dirm}{when two faces are plotted, what should connect the corresponding markers}

\item{subset}{of markers to be shown}
}
\value{
nothing - a plot is constructed
}
\description{
Plot a face or a pair of faces
}
