% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{read_face_file}
\alias{read_face_file}
\title{Read in a face file}
\usage{
read_face_file(fname, mfconfig, numskip = 5)
}
\arguments{
\item{fname}{The name of the function}

\item{mfconfig}{motion file configuration}

\item{numskip}{The number of lines to skip at the beginning of the motion file}
}
\value{
a data matrix of the motion file
}
\description{
Read in a face file
}
