% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{readlmk}
\alias{readlmk}
\title{Read in an LMK file}
\usage{
readlmk(filename)
}
\arguments{
\item{filename}{- name of the file}
}
\value{
list containing the short labels, long labels and matrix of landmarks
}
\description{
Read in an LMK file
}
