% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{reflectface}
\alias{reflectface}
\title{Reflect face}
\usage{
reflectface(sm, mfconfig)
}
\arguments{
\item{sm}{face matrix}

\item{mfconfig}{face config info including left right pairs}
}
\value{
a face matrix
}
\description{
Reflect face by multiplying lateral direction by -1 and swapping the tags
of paired markers
}
\details{
This function will work best when the face is centered with zero on
the midplane (lateral direction is tangent to this). But this is not
essential as reflection will happen anyway although the face may need to
be translated and/or rotated to match it up with the original face.
}
