% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{rotface}
\alias{rotface}
\title{Rotate a face matrix in SD}
\usage{
rotface(sm, theta, dirs)
}
\arguments{
\item{sm}{The face matrix}

\item{theta}{Angle of rotation}

\item{dirs}{Coordinate axes of plane of rotation - first dir is theta=0}
}
\value{
a face matrix
}
\description{
Rotate a face matrix in SD
}

