% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{trajtemplate}
\alias{trajtemplate}
\title{Conform a trajectory to a template}
\usage{
trajtemplate(sm, mfi)
}
\arguments{
\item{sm}{The motion data array (may contain missing values)}

\item{mfi}{template shape (no missing values allowed)}
}
\value{
a list with the following components:
\item{sm}{conformed trajectory}
\item{dm}{matrix of squared distances to the template by frame and by marker}
\item{rawstart}{the initial frame}
\item{rawext}{the extreme frame}
}
\description{
Conform a trajectory to a template
}
