% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facer.R
\name{facemovie}
\alias{facemovie}
\title{Make a facial motion movie}
\usage{
facemovie(
  shapem1,
  mfconfig,
  shapem2 = NULL,
  everynth = 6,
  adjust = TRUE,
  moviename = "movie.mov",
  header = NULL,
  missmat = NULL
)
}
\arguments{
\item{shapem1}{A shape motion}

\item{mfconfig}{motion config}

\item{shapem2}{another shape motion (optional)}

\item{everynth}{sample every nth frame (default is 6)}

\item{adjust}{adjust second motion to start from same place as first motion}

\item{moviename}{name of the movie (default is "movie.mov")}

\item{header}{title displayed on movie}

\item{missmat}{0/1 nframes by nmarkers matrix with 1's for missing values}
}
\value{
nothing - find movie file in "animations" directory. Second sequence is in red
}
\description{
Requires availability of the convert function from Imagemagick
}
